requirements.txt
dgl==0.6.1
networkx==2.2
numpy==1.19.5
pandas==0.25.2
pytorch-lightning==1.2.5
pytorch-lightning-bolts==0.3.0
sklearn==0.0